ecocivic-dapp/backend-ai/requirements.txt
flask
flask-cors
python-dotenv
pytesseract
pillow
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
gunicorn>=21.2.0
flask-limiter>=3.5.0
PyJWT>=2.8.0
python-dateutil>=2.8.2
Flask-APScheduler>=1.12.4
web3>=6.0.0